More aggressive mutex locking on connections
The connections for a given ip address dictionary was being overwritten
when new connections occurred. This created deadlock scenarios in which
consensus could not occur.

It originally worked in small networks but as we scaled and gained more
connections issues arose. The solution was to lock before checking if a
key was in the dictionary and then only release after updating the
dictionary or checking if the key existed, thus preventing race
conditions. I also had to include checks before every time a new
connection was added because we didn't want to have multiple connections
for each address - only a single one.